sample_code/s3-car-manual-search/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
boto3>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0